This is the main method
The main reason I have created this project is to let the user enter and store a recipe, it can only store one recipe so far. The main method contains other methods that are called from the external class named Functions. In the main method I used a switch case to let or prompt the user to choose the options between the menu. the menu also contains color to bring out the creative